WIP on fixing tests.
